🔧schedule page 관련 container 분리 작업
